# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive, modular MCP (Model Context Protocol) proxy system",
        "title": "SUSE AI Universal Proxy API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "SUSE",
            "url": "https://github.com/suse/suse-ai-up",
            "email": "info@suse.ai"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8911",
    "basePath": "/",
    "paths": {
...
        "/api/v1/auth/oauth/callback": {
            "post": {
                "description": "Handle OAuth provider callback",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "OAuth callback",
                "parameters": [
                    {
                        "description": "OAuth callback data",
                        "name": "callback",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.OAuthCallbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/discovery/servers/{id}": {
            "get": {
                "description": "This endpoint is deprecated due to data access issues. Use GET /api/v1/discovery/results/{id} instead.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "discovery"
                ],
                "summary": "Get discovered server by ID (DEPRECATED)",
                "deprecated": true,
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DiscoveredServer"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "handlers.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "$ref": "#/definitions/models.AuthToken"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                }
            }
        },
        "handlers.OAuthCallbackRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
...
        "models.AuthAnalysis": {
            "type": "object",
            "properties": {
                "confidence": {
                    "description": "\"high\", \"medium\", \"low\"",
                    "type": "string"
                },
                "detected_mechanisms": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "required": {
                    "type": "boolean"
                },
                "type": {
                    "description": "\"none\", \"basic\", \"bearer\", \"apikey\", \"oauth\", \"missing\"",
                    "type": "string"
                },
                "vulnerabilities": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.AuthToken": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "token_type": {
                    "description": "\"Bearer\"",
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "models.ConnectionType": {
            "type": "string",
            "enum": [
                "SSE",
                "StreamableHttp",
                "RemoteHttp",
                "LocalStdio",
                "SidecarStdio"
            ],
            "x-enum-varnames": [
                "ConnectionTypeSSE",
                "ConnectionTypeStreamableHttp",
                "ConnectionTypeRemoteHttp",
                "ConnectionTypeLocalStdio",
                "ConnectionTypeSidecarStdio"
            ]
        },
        "models.DiscoveredServer": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "http://192.168.1.100:8000"
                },
                "auth_info": {
                    "$ref": "#/definitions/models.AuthAnalysis"
                },
                "capabilities": {
                    "description": "Enhanced fields for deep interrogation",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.McpCapabilities"
                        }
                    ]
                },
                "connection": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ConnectionType"
                        }
                    ],
                    "example": "StreamableHttp"
                },
                "id": {
                    "type": "string",
                    "example": "server-123"
                },
                "lastSeen": {
                    "type": "string"
                },
                "last_deep_scan": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "MCP Example Server"
                },
                "prompts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.McpPrompt"
                    }
                },
                "protocol": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ServerProtocol"
                        }
                    ],
                    "example": "MCP"
                },
                "protocol_version": {
                    "type": "string"
                },
                "resource_templates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.McpResourceTemplate"
                    }
                },
                "resources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.McpResource"
                    }
                },
                "server_version": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "healthy"
                },
                "tools": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.McpTool"
                    }
                },
                "vulnerability_score": {
                    "type": "string",
                    "example": "high"
                }
            }
        },
...
        "models.McpArgument": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "required": {
                    "type": "boolean"
                }
            }
        },
        "models.McpCapabilities": {
            "type": "object",
            "properties": {
                "completions": {
                    "type": "boolean"
                },
                "experimental": {
                    "type": "boolean"
                },
                "logging": {
                    "type": "boolean"
                },
                "prompts": {
                    "type": "boolean"
                },
                "resources": {
                    "type": "boolean"
                },
                "tools": {
                    "type": "boolean"
                }
            }
        },
        "models.McpPrompt": {
            "type": "object",
            "properties": {
                "arguments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.McpArgument"
                    }
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.McpResource": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "mime_type": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "uri": {
                    "type": "string"
                }
            }
        },
        "models.McpResourceTemplate": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "mime_type": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "uri_template": {
                    "type": "string"
                }
            }
        },
        "models.McpTool": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "input_schema": {
                    "type": "object",
                    "additionalProperties": true
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.ServerProtocol": {
            "type": "string",
            "enum": [
                "MCP"
            ],
            "x-enum-varnames": [
                "ServerProtocolMCP"
            ]
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "auth_provider": {
                    "type": "string",
                    "example": "local"
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "external_id": {
                    "type": "string",
                    "example": "github123"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"mcp-users\"",
                        "\"weather-team\"]"
                    ]
                },
                "id": {
                    "type": "string",
                    "example": "user123"
                },
                "last_login_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "password_changed_at": {
                    "type": "string"
                },
                "provider_groups": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"org/team\"]"
                    ]
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}